+++ src/tiposDeDatos/Entero.java
@@ -6,7 +6,6 @@ import operadores.OperadorUnario;
 
 public class Entero implements TipoDeDato<Integer> {
 
-
     @Override
     public Integer parseDato(String stringDato) {
         try {
@@ -18,12 +17,12 @@ public class Entero implements TipoDeDato<Integer> {
 
     @Override
     public Integer operar(Operador op, Object op1, Object op2) throws Exception  {
-        return op.operar((Integer) op1,(Integer)op2);
+        return (Integer)op.operar((Integer) op1,(Integer)op2);
     }
 
     @Override
     public Integer operar(OperadorUnario op, Object op1) throws Exception {
-        return op.operar((Integer)op1);
+        return (Integer)op.operar((Integer)op1);
     }
 
     @Override
